+++ protractor.conf.js
@@ -25,12 +25,12 @@ exports.config = {
     cucumberOpts: {
         // Require files before executing the features.
         require: [
-            'features/step_definitions/**/*.js',
+            'features/components/**/*.js',
             'features/support/**/*.js'
         ],
         // Only execute the features or scenarios with tags matching @dev.
         // This may be an array of strings to specify multiple tags to include.
-        // tags: '@dev',
+        tags: ['~@ignore','~@wip'],
         // How to format features (progress, summary, pretty, json)
         format: ['progress', 'pretty', 'json:test_reports/cucumber_report.json'],
     },