+++ RELEASE_NOTES.md
@@ -10,6 +10,10 @@ _Date TBD_
 * Displayed filtered Test Standards on Criteria editing
 * Added GTM tracking to Search - Criteria/CQMs/Surveillance filters
 
+### Text edits
+* Added description to Nonconformities page
+* Added note about 2011 Decertified Products
+
 =======
 
 ## Version 8.3.0