+++ RELEASE_NOTES.md
@@ -9,6 +9,11 @@ _Date TBD_
   * Display errors if Listing was previously "completed"
 * Displayed filtered Test Standards on Criteria editing
 * Added GTM tracking to Search - Criteria/CQMs/Surveillance filters
+* Added "Questionable Activity Report"
+
+### Text edits
+* Added description to Nonconformities page
+* Added note about 2011 Decertified Products
 
 =======